+++ tests/integration/test_subscription.py
@@ -524,6 +524,37 @@ class TestSubscription(unittest.TestCase):
         self.assertEqual(1, len(phone_errors))
         self.assertEqual(ErrorCodes.Descriptor.PhoneFormatIsInvalid, phone_errors[0].code)
 
+    def test_description(self):
+        http = ClientApiHttp.create()
+        status_code, nonce = http.get_paypal_nonce({
+            "consent-code": "consent-code",
+            "options": {"validate": False}
+        })
+        self.assertEqual(202, status_code)
+
+        payment_method_token = PaymentMethod.create({
+            "customer_id": Customer.create().customer.id,
+            "payment_method_nonce": nonce
+        }).payment_method.token
+
+        result = Subscription.create({
+            "payment_method_token": payment_method_token,
+            "plan_id": TestHelper.trialless_plan["id"],
+            "options": {
+                "paypal": {
+                    "description": "A great product"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        subscription = result.subscription
+        self.assertEqual("A great product", subscription.description)
+
+        transaction = subscription.transactions[0]
+        paypal_details = transaction.paypal_details
+        self.assertEqual("A great product", paypal_details.description)
+
     def test_find_with_valid_id(self):
         subscription = Subscription.create({
             "payment_method_token": self.credit_card.token,
@@ -949,6 +980,41 @@ class TestSubscription(unittest.TestCase):
 
         self.assertEqual("999*99", updated_subscription.descriptor.name)
         self.assertEqual("1234567890", updated_subscription.descriptor.phone)
+
+    def test_update_description(self):
+        http = ClientApiHttp.create()
+        status_code, nonce = http.get_paypal_nonce({
+            "consent-code": "consent-code",
+            "options": {"validate": False}
+        })
+        self.assertEqual(202, status_code)
+
+        payment_method_token = PaymentMethod.create({
+            "customer_id": Customer.create().customer.id,
+            "payment_method_nonce": nonce
+        }).payment_method.token
+
+        result = Subscription.create({
+            "payment_method_token": payment_method_token,
+            "plan_id": TestHelper.trialless_plan["id"],
+            "options": {
+                "paypal": {
+                    "description": "A great product"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        subscription = result.subscription
+        updated_subscription = Subscription.update(subscription.id, {
+            "options": {
+                "paypal": {
+                    "description": "An incredible product"
+                }
+            }
+        }).subscription
+
+        self.assertEqual("An incredible product", updated_subscription.description)
 
     def test_cancel_with_successful_response(self):
         subscription = Subscription.create({