+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Coinbase is no longer a supported payment method. `PaymentMethodNoLongerSupported` will be returned for Coinbase operations.
 * Add facilitated details to Transaction if present
 * Add `submit_for_settlement` option to `Subscription.retry_charge`
+* Add `options` -> `paypal` -> `description` for creating and updating subscriptions
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`